[dependency refactor] Refactor OpenAI wrapper
Summary:

Test Plan:
